backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
scikit-learn
pydantic-settings
python-dotenv
openpyxl